chore: move custom files from customizations/ to customizations/src/
