+++ packages/core/next.config.js
@@ -16,7 +16,7 @@ const nextConfig = {
     defaultLocale: 'en-US',
   },
   sassOptions: {
-    additionalData: `@import "src/customizations/styles/custom-mixins.scss";`,
+    additionalData: `@import "src/customizations/src/styles/custom-mixins.scss";`,
   },
   // TODO: We won't need to enable this experimental feature when migrating to Next.js 13
   experimental: {
