+++ packages/core/src/components/cms/GlobalSections.tsx
@@ -1,7 +1,7 @@
 import { Locator, Section } from '@vtex/client-cms'
 import type { ComponentType } from 'react'
 import { PropsWithChildren, lazy } from 'react'
-import CUSTOM_COMPONENTS from 'src/customizations/components'
+import CUSTOM_COMPONENTS from 'src/customizations/src/components'
 import { PageContentType, getPage } from 'src/server/cms'
 
 import Toast from 'src/components/common/Toast'
