+++ packages/core/src/components/sections/Alert/Overrides.tsx
@@ -1,7 +1,7 @@
 import { Alert as UIAlert, Icon as UIIcon } from '@faststore/ui'
 
 import { getSectionOverrides } from 'src/utils/overrides'
-import { override } from 'src/customizations/components/overrides/Alert'
+import { override } from 'src/customizations/src/components/overrides/Alert'
 import type { AlertOverrideDefinition } from 'src/typings/overrides'
 
 const { Alert, Icon } = getSectionOverrides(
