+++ packages/core/src/components/sections/EmptyState/Overrides.tsx
@@ -1,7 +1,7 @@
 import { EmptyState as UIEmptyState } from '@faststore/ui'
 
 import { getSectionOverrides } from 'src/utils/overrides'
-import { override } from 'src/customizations/components/overrides/EmptyState'
+import { override } from 'src/customizations/src/components/overrides/EmptyState'
 import type { EmptyStateOverrideDefinition } from 'src/typings/overrides'
 
 const { EmptyState } = getSectionOverrides(
