+++ packages/core/src/components/sections/ProductShelf/Overrides.tsx
@@ -4,7 +4,7 @@ import ProductCard from 'src/components/product/ProductCard'
 import Carousel from 'src/components/ui/Carousel'
 
 import { getSectionOverrides } from 'src/utils/overrides'
-import { override } from 'src/customizations/components/overrides/ProductShelf'
+import { override } from 'src/customizations/src/components/overrides/ProductShelf'
 import type { ProductShelfOverrideDefinition } from 'src/typings/overrides'
 
 const { ProductShelf, __experimentalCarousel, __experimentalProductCard } =
