+++ packages/core/src/components/sections/RegionBar/Overrides.tsx
@@ -1,7 +1,7 @@
 import { RegionBar as UIRegionBar, Icon as UIIcon } from '@faststore/ui'
 
 import { getSectionOverrides } from 'src/utils/overrides'
-import { override } from 'src/customizations/components/overrides/RegionBar'
+import { override } from 'src/customizations/src/components/overrides/RegionBar'
 import type { RegionBarOverrideDefinition } from 'src/typings/overrides'
 
 const { RegionBar, LocationIcon, ButtonIcon } = getSectionOverrides(
