+++ packages/core/src/components/templates/SearchPage/SearchPage.tsx
@@ -4,7 +4,7 @@ import type { ComponentType } from 'react'
 import Breadcrumb from 'src/components/sections/Breadcrumb'
 import ProductGallery from 'src/components/sections/ProductGallery'
 import { ITEMS_PER_PAGE } from 'src/constants'
-import CUSTOM_COMPONENTS from 'src/customizations/components'
+import CUSTOM_COMPONENTS from 'src/customizations/src/components'
 import RenderSections from 'src/components/cms/RenderSections'
 import { SearchContentType } from 'src/server/cms'
 import { useProductGalleryQuery } from 'src/sdk/product/useProductGalleryQuery'
