--- packages/core/src/customizations/components/overrides/ThirdPartyScripts.tsx
@@ -1,9 +1,0 @@
-import { default as ThirdPartyScripts } from 'src/customizations/scripts/ThirdPartyScripts'
-
-const overrides = {
-  components: {
-    ThirdPartyScripts,
-  },
-}
-
-export default overrides
