+++ packages/core/src/customizations/src/GlobalOverrides.tsx
@@ -1,6 +1,6 @@
-import WebFontsOverrides from 'src/customizations/components/overrides/WebFonts'
+import WebFontsOverrides from 'src/customizations/src/components/overrides/WebFonts'
 import { default as CoreWebFonts } from 'src/fonts/WebFonts'
-import ThirdPartyScriptsOverrides from 'src/customizations/components/overrides/ThirdPartyScripts'
+import ThirdPartyScriptsOverrides from 'src/customizations/src/components/overrides/ThirdPartyScripts'
 
 const Components = {
   WebFonts: CoreWebFonts,
