+++ packages/core/src/customizations/src/components/CustomComponent.tsx
