+++ packages/core/src/customizations/src/components/overrides/BannerText.tsx
