+++ packages/core/src/customizations/src/components/overrides/Hero.tsx
