+++ packages/core/src/customizations/src/components/overrides/ProductDetails.tsx
