+++ packages/core/src/customizations/src/components/overrides/ProductGallery.tsx
