+++ packages/core/src/customizations/src/components/overrides/ThirdPartyScripts.tsx
@@ -1,0 +1,9 @@
+import { default as ThirdPartyScripts } from 'src/customizations/src/scripts/ThirdPartyScripts'
+
+const overrides = {
+  components: {
+    ThirdPartyScripts,
+  },
+}
+
+export default overrides
