+++ packages/core/src/customizations/src/components/overrides/WebFonts.tsx
@@ -1,4 +1,4 @@
-import { default as CustomWebFonts } from 'src/customizations/fonts/WebFonts'
+import { default as CustomWebFonts } from 'src/customizations/src/fonts/WebFonts'
 
 const overrides = {
   components: {
