+++ packages/core/src/customizations/src/fragments/ClientManyProducts.ts
