+++ packages/core/src/customizations/src/fragments/ClientProduct.ts
