+++ packages/core/src/customizations/src/fragments/ClientProductGallery.ts
