+++ packages/core/src/customizations/src/fragments/ClientSearchSuggestions.ts
