+++ packages/core/src/customizations/src/fragments/ClientTopSearchSuggestions.ts
