+++ packages/core/src/customizations/src/fragments/ServerProduct.ts
