+++ packages/core/src/customizations/src/scripts/ThirdPartyScripts.tsx
