+++ packages/core/src/pages/_app.tsx
@@ -1,7 +1,7 @@
 // FastStore UI's base styles
 import '../styles/global/index.scss'
 
-import '../customizations/themes/index.scss'
+import '../customizations/src/themes/index.scss'
 
 import type { AppProps } from 'next/app'
 import NextNProgress from 'nextjs-progressbar'
