+++ packages/core/src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import { Head, Html, Main, NextScript } from 'next/document'
 
 import ThirdPartyScripts from 'src/components/ThirdPartyScripts'
-import { WebFonts } from 'src/customizations/GlobalOverrides'
+import { WebFonts } from 'src/customizations/src/GlobalOverrides'
 
 function Document() {
   return (
